refactor: Phase 3 - reorganize project commands for clear separation
- Rename commands/ to project-commands/ for clarity
- Create project-commands/index.ts for clean exports
- Update import paths in src/app/index.ts
- Establish clear distinction:
  * project-commands/ = saiki init, create-app (run once, exit)
  * interactive-commands/ = /session, /model, /mcp (CLI session)
- All 541 tests pass, build successful, no regression

Project setup now clearly separated from interactive CLI commands.
