+++ src/app/cli/project-commands/index.ts
@@ -1,0 +1,18 @@
+// Project setup commands - run once and exit
+export {
+    createSaikiProject,
+    createTsconfigJson,
+    addSaikiScriptsToPackageJson,
+    postCreateSaiki,
+} from './create.js';
+
+export {
+    getUserInputToInitSaikiApp,
+    initSaiki,
+    postInitSaiki,
+    createSaikiDirectories,
+    createSaikiConfigFile,
+    updateSaikiConfigFile,
+    createSaikiExampleFile,
+    updateEnvFile,
+} from './init.js';
